Refactor for clarity
Attempted to merge the FilePrinter and FilePathPrompt classes because
they seemed at first look to be working on the same task. Was not able
to replicate the behavior I wanted, and realized that the path must be
validated before entering the try block for printing in FilePrinter in
order to maintain the current error-to-user structure.