Change coin message codes
Because the program must determine which coin specification to use based
on what is entered, all of the message files must contain information
for all possible currencies. It would be easier to inject arrays
specifying the currencies into different instances of the coin
calculator to allow them to be easily configured according to location,
but this would not meet the command-line argument currency requirement.