Convert null checks to exception throw and catch
Move all exception handling to the top of the class to isolate it from
the rest of the code. Change null checks to throw exceptions. Create a
new exception type to cover the case of a null value being returned from
the dao. Catching any NullPointerException within the method that calls
the dao obfuscated potential issues not related to the dao returning a
null value.